Refactor graphql module
Closes: #80
